fix: change getVoltage to explicitly allow communication via I2C
At the moment, if you want to take a voltage reading, you cannot do so if there is an external sensor connected to the Margay. This update to the getVoltage function first tests whether the external I2C connection is on. If it is, then it takes a reading, if it is not, then it turns it on, takes a reading, and then turns it back off again.