feat: add farm-gate principle to I2C bus switching
add logic to check external I2C bus status before making a to open or close the bus.
Add new function to close gate if closed, and keep gate open if was open.
This works by checking the current state, saving it to a local variable, then comparing to a
class-scope variable that preserves the current state of the system.